requirements.txt

netmiko>=4.0.0
napalm
pyyaml
Jinja2
typer
